Split show notes tests by scenario
Move shared show-notes LLM test helpers into a support module, then
split parsing, prompting, generation, and TEI enrichment coverage into
focused test modules.

Remove the deleted monolithic test module and its file-level
`too-many-lines` suppression.
